AOP & DI - using AspectJ-Style Annotation
