+++ myapp/simple-console-app/src/main/java/com/learn/simpleconsoleapp/beans/GreatSinger.java
@@ -1,10 +1,12 @@
 package com.learn.simpleconsoleapp.beans;
 
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
 @Component
+@Qualifier("greatSinger")
 @Trophy(name = {"grammy", "diamond disk"})
-public class GreatSinger extends Singer {
+public class GreatSinger extends SimpleSinger {
     private String lyric = "We found a message in a bottle we were drinking";
     protected float incomePercent = 0.9f;
