+++ myapp/simple-console-app/src/main/java/com/learn/simpleconsoleapp/configs/demo/SingerDocumentaryUsingAspectjConfig.java
@@ -1,0 +1,15 @@
+package com.learn.simpleconsoleapp.configs.demo;
+
+import org.springframework.context.annotation.*;
+
+/**
+ * This config can not exist with other config -> cause of conflict @EnableAspectJAutoProxy
+ */
+
+//@Configuration
+//@ComponentScan(basePackages = {"com.learn.simpleconsoleapp"},
+//        excludeFilters = {@ComponentScan.Filter(type = FilterType.REGEX, pattern = {"com.learn.simpleconsoleapp.configs.*"})})
+//@PropertySource({"classpath:application.properties"})
+//@EnableAspectJAutoProxy()
+//public class SingerDocumentaryUsingAspectjConfig {
+//}
