+++ myapp/simple-console-app/src/main/resources/application.properties
@@ -1,2 +1,3 @@
 my.name=Thomas
 my.city=HCM
+aspect.console.logging.singer.sing=blah
