+++ myapp/simple-console-app/src/main/resources/demo/singer-documentary-using-aspectj-context.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns:context="http://www.springframework.org/schema/context"
+       xmlns:aop="http://www.springframework.org/schema/aop"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd http://www.springframework.org/schema/context https://www.springframework.org/schema/context/spring-context.xsd http://www.springframework.org/schema/aop https://www.springframework.org/schema/aop/spring-aop.xsd">
+
+    <context:component-scan base-package="com.learn.dummybeans,com.learn.moneytransfer,com.learn.simpleconsoleapp">
+        <context:exclude-filter type="regex" expression="com.learn.simpleconsoleapp.configs.demo.*"/>
+    </context:component-scan>
+
+    <context:property-placeholder location="classpath:application.properties,classpath:msf.properties"/>
+
+    <aop:aspectj-autoproxy/>
+
+</beans>
